requirements.txt
kafka-python
influxdb-client
